refactor: create integration_tests crate
