+++ Cargo.toml
@@ -10,7 +10,7 @@ members = [
     "dan_layer/template_lib",
     "dan_layer/template_macros",
     "dan_layer/template_test_tooling",
-    "dan_layer/tests",
+    "dan_layer/integration_tests",
     "applications/tari_validator_node",
     "applications/tari_validator_node_cli",
 ]
