+++ dan_layer/integration_tests/Cargo.toml
@@ -4,20 +4,19 @@ version = "0.35.1"
 edition = "2018"
 license = "BSD-3-Clause"
 
-[dev-dependencies]
+[dependencies]
 tari_common = { git = "https://github.com/tari-project/tari.git", branch = "feature-dan", package = "tari_common", features = ["build"] }
 tari_common_types = { git = "https://github.com/tari-project/tari.git", branch = "feature-dan", package = "tari_common_types" }
 tari_crypto = { git = "https://github.com/tari-project/tari-crypto.git", tag = "v0.15.6" }
 tari_dan_common_types = { path = "../common_types" }
 tari_dan_core = { path = "../core" }
 tari_dan_engine = { path = "../engine" }
-tari_dan_storage_sqlite = { path = "../storage_sqlite", features=["test"]}
+tari_dan_storage_sqlite = { path = "../storage_sqlite" }
 tari_engine_types = { path = "../engine_types" }
 tari_shutdown = { git = "https://github.com/tari-project/tari.git", branch = "feature-dan", package = "tari_shutdown" }
 tari_template_lib = { path = "../template_lib" }
 tari_test_utils = { git = "https://github.com/tari-project/tari.git", branch = "feature-dan", package = "tari_test_utils" }
 tari_utilities = { git = "https://github.com/tari-project/tari_utilities.git", tag = "v0.4.7" }
-
 
 lazy_static = "1.4.0"
 rand = "0.7"
